+++ prow/plank/controller.go
@@ -30,11 +30,11 @@ import (
 	prowapi "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	"k8s.io/test-infra/prow/config"
 	"k8s.io/test-infra/prow/github"
+	reportlib "k8s.io/test-infra/prow/github/report"
 	"k8s.io/test-infra/prow/github/reporter"
 	"k8s.io/test-infra/prow/kube"
 	"k8s.io/test-infra/prow/pjutil"
 	"k8s.io/test-infra/prow/pod-utils/decorate"
-	reportlib "k8s.io/test-infra/prow/report"
 )
 
 const (
